Add files via upload
These are the files to Car Dodger. The game was built on a Windows 10 machine, 8 gb of ram, 512 mb of vram soldered into the CPU(accelerated AMD  APU @ approximately 1.8 gHz). Has not been tested on previous operating systems,email me at hezekiah.s2000@gmail.com if you are able to do so on a Windows 7 machine and list the specifications of your computer as well, it would much appreciated,thank you.